arquivos/requirements.txt
Flask
mysql-connector-python
bcrypt
#pip install -m requeriments.txt